Support for on-the-fly change of the fairness definition build in
The timed fairness might not work correctly yet (not fully tested and no time to do so right now),and it slows down the simulator a lot. The framework is done though (some changes needed on the long run) and we can show that on-the-fly switching of the definitions works.